Fix OSM commands for the western hemisphere
The OSMcommand library was unintentionally bounding the minimum at [0,0].
When the location for the OSM command (such as 'rails') was in the western
hemisphere (ie, a negative number), the calculation for the 'minimum'
coordinates would be improperly calculating, leading to a faulty request
to the OpenStreetMap API, which would respond with a 400 error.

Fix it by bounding the minimum value at the most negative number.

Fixes: #10